prov/sm2: Add FI_ATOMIC support to SM2
This commit adds FI_ATOMIC support to the SM2 provider. The SM2 atomic
code was based of the SHM atomic implementation, and modified in two
major ways.

The first big modification from SHM in SM2 is that we stick the entire
atomic operation into the 4k inject buffer. This includes the rma ioc,
result iov, compare iov and iov data. We use a series of structures to
define what the user_data of the xfer entry looks like, and we cast the
user_data to be of type sm2_atomic_entry.

The second big modification from SHM in SM2 is that we use the built-in
buffer return protocol to do our atomic operations which previously required
READ RMA semantics.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>