prov/efa: Make rxe_map part of the efa_rdm_peer struct
This change is necessary to implement an implicit AV.
With an implicit AV, using the fi_addr is no longer an option for
identifying different peers. All unknown peers will have fi_addr set
to -1. And the index in the implicit AV can be re-used, so it's not
guaranteed to be unique.

This commit makes the rxe_map part of the peer struct and only uses the
msg_id as the hash key

Signed-off-by: Sai Sunku <sunkusa@amazon.com>