fabtests/efa: Fix the av operation
Currently, open_client uses an intermediate ptr av
to stage the value of avs array element, and only open
av for that av ptr only. However, this av is not tracked
after the function returns and become a wild pointer.

This patch fixes this issue by operating on the avs array
directly so it can be cleaned up properly during the free
functions

Signed-off-by: Shi Jin <sjina@amazon.com>