prov/efa: Don't give up the srx lock between wait send and destroying QP
Before this commit, the endpoint close path acquired the SRX lock in two
different stages: first for the wait send to finish and second to
destroy the QP

But an fi_cq_read operation can happen in between and cause illegal
memory accesses. This commit keeps the srx lock for the entire endpoint
close procedure.

Signed-off-by: Sai Sunku <sunkusa@amazon.com>